F [#TSS-10] Separate organization out of the connector class
